Add user authentication and authorization to Todo List API
- Implement user registration and login functionality
- Secure routes with authentication middleware
- Associate todo lists with users in the database
- Validate user input with Joi schemas

+++ controllers/todoList-controller.js
@@ -3,14 +3,19 @@ import { todoListSchema } from "../schemas/todoList-schema.js"
 
 export const createList = async(req,res) =>{
     try {
+        const userID = req.user.id
         // validate with joi
         const {error,value} = todoListSchema.validate(req.body);
         if(error) {
             return res.status(400).json({message:error.details[0].message})
         }
+        if (userID.toString() !== value.user.toString()){
+            return res.status(403).json({message:'You are not authorized to create a list for this user'})
+        }
         // create your Todo list
         const createTodoListData = await TodoList.create(value)
-        return res.status(200).json({message:'Your List is Successfully created🎉'})
+        console.log('New list', createTodoListData)
+        return res.status(202).json({message:'Your List is Successfully created🎉'})
 
     } catch (error) {
         return res.status(500).json({message:error.message})
@@ -20,16 +25,14 @@ export const createList = async(req,res) =>{
 
 export const lists = async(req,res) =>{
     try {
-        // find all List created
-        const allList = await TodoList.find();
+        const userID = req.user.id
+        // find all List created by the user
+        const allLists = await TodoList.find({user:userID});
         // return a message if there is no list
-        if(allList == ''){
+        if(allLists.length == 0){
             return res.status(400).json({message:'You have no list'})
         };
-        return res.status(200).json(
-            {message:`These are all your List`},
-            allList
-        );
+        return res.status(200).json({message:`These are all your List`, lists: allLists });
     } catch (error) {
         return res.status(500).json({message:error.message})
     };
@@ -39,6 +42,7 @@ export const singleList = async(req, res) =>{
     try {
         // check the id 
         const listID = req.params.id;
+        const userID = req.user.id;
         if(!listID){
             return res.status(400).json({message:`List ID: ${listID}, not available`});
         }
@@ -47,10 +51,11 @@ export const singleList = async(req, res) =>{
         if(!list){
             return res.status(400).json({message:'Invalid List'})
         }
-        return res.status(200).json(
-            {message:'List is Found'},
-            list
-        );
+        // check if list belong to the user
+        if(list.user.toString() !== userID.toString()){
+            return res.status(403).json({message:'You are not authorized to see a list for this user'})
+        }
+        return res.status(200).json({message:'List is Found', singleList: list},);
     } catch (error) {
         return res.status(500).json({message: error.message}) 
     };
@@ -58,6 +63,7 @@ export const singleList = async(req, res) =>{
 
 export const updateList = async(req, res) =>{
     try {
+        const userID = req.user.id;
         // check the id 
         const listID = req.params.id;
         if(!listID){
@@ -68,15 +74,17 @@ export const updateList = async(req, res) =>{
         if(!list){
             return res.status(400).json({message:'Invalid List'})
         }
+        // check if list belong to the user
+        if(list.user.toString() !== userID.toString()){
+            return res.status(403).json({message:'You are not authorized to see a list for this user'})
+        }
         // upadate a list
         const newList = await TodoList.findByIdAndUpdate(
+            listID,
             req.body,
             {new:true}
         );
-        return res.status(200).json(
-            {message:'List is updated Successfully'},
-            newList
-        );
+        return res.status(200).json({message:'List is updated Successfully', udateList: newList});
     } catch (error) {
         return res.status(500).json({message: error.message}) 
     };
@@ -84,6 +92,7 @@ export const updateList = async(req, res) =>{
 
 export const deleteList = async(req, res) =>{
     try {
+        const userID = req.user.id;
         // check the id 
         const listID = req.params.id;
         if(!listID){
@@ -94,12 +103,13 @@ export const deleteList = async(req, res) =>{
         if(!list){
             return res.status(400).json({message:'Invalid List'})
         }
-        // upadate a list
+        // check if list belong to the user
+        if(list.user.toString() !== userID.toString()){
+            return res.status(403).json({message:'You are not authorized to see a list for this user'})
+        }
+        // delete a list
         const newList = await TodoList.findByIdAndDelete(listID);
-        return res.status(200).json(
-            {message:'List is Deleted Successfully'},
-            newList
-        );
+        return res.status(200).json({message:'List is Deleted Successfully'});
     } catch (error) {
         return res.status(500).json({message: error.message}) 
     };
